chore(stories): organize stories
